Move BowlingCenterQueries to Database/Queries
Refactor: relocate BowlingCenterQueries into src/Neba.Infrastructure/Database/Queries and update namespaces/usings accordingly. Updated DatabaseConfiguration to reference the new Neba.Infrastructure.Database.Queries namespace and removed an unused Neba.Infrastructure.Database using. Renamed/moved corresponding test files and adjusted their namespaces/usings to tests/Neba.Infrastructure.Tests/Database/Queries so tests reflect the new location.
