+++ tests/Neba.Infrastructure.Tests/Database/Queries/BowlingCenterQueriesTests.GetAllAsync_ReturnsOnlyOpenBowlingCenters.verified.txt
